+++ NEWS.md
@@ -1,3 +1,9 @@
+# kerntools 1.1.0
+
+* New vignettes explaining in depth kernel PCA and the kernel functions implemented in this package.
+
+* New example dataset ('soil') that contains bacterial counts.
+
 # kerntools 1.0.2
 
 * Fixed errors that arose during CRAN Package Check when using alternative BLAS/LAPACK implementations.